pyproject.toml
[project]
name = "unifi-release-announcer"
version = "0.1.0"
description = "Discord bot that announces new UniFi Protect releases"
authors = [
    {name = "damacus"}
]
requires-python = ">=3.11"
dependencies = [
    "discord.py",
    "playwright>=1.55.0",
    "pynacl>=1.5.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "defusedxml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
    "types-defusedxml",
    "respx",
    "pytest-asyncio",
]

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = []

[tool.setuptools.packages.find]
where = ["./"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
